Allow constant write enable signals to optimize
After encountering a timing error associated with RAM write enable signals
in an FFT, it emerged that a constant we signal, which went through a series of
pipeline registers and then was fanned out to multiple brams, was not being
optimized away properly.

This commit adds a parameter to the bus_single_port_ram block, to allow the
user to explicitly state that the write enable signals are constantly high.
In this case, the write enable fanout registering and logic is replaced with a
single constant block. This change has been propagated to the fft_stage_n block.

The we input to the bus_single_port_ram block is maintained for backwards compatibility
even if it is not used. The default behaviour of the block is to assume this we input is
not constant.