Fixed FFT fanout latency bug
When delays in the fft_stage_n block are implemented in slices
there is no fanout delay associated with registering bram address lines.
However, the initialisation script was compensating for this non-existant delay
by adding latency on the sync lines. This commit fixes this, by catching the case
where there will be no fanouts and setting fanout_latency to 0.