Default to truncate/wrap if full precision addsub
Previously, the mask parameters defining the quantization
and overflow strategy would be used, even if the addsub block
was set to full-precision output. I noticed this bug in the FFT,
where the resulting logic ends up causing large fanout signals
in the rescale by 2^-1 mux.

This commit turns the cast strategy to truncate + wrap when
full output precision is selected or (as is the case in the FFT)
when the output precision is manually specified but amounts to
full precision.